+++ publiccms-by-gradle/src/com/publiccms/controller/admin/sys/SysDomainAdminController.java
@@ -1,9 +1,10 @@
 package com.publiccms.controller.admin.sys;
 
+import static com.sanluan.common.tools.JsonUtils.getString;
+
 // Generated 2016-1-27 21:06:08 by com.sanluan.common.source.SourceMaker
 
 import static com.sanluan.common.tools.RequestUtils.getIpAddress;
-import static com.sanluan.common.tools.JsonUtils.getString;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
@@ -12,6 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.publiccms.common.base.AbstractController;
 import com.publiccms.entities.log.LogOperate;
@@ -25,18 +27,18 @@ import com.publiccms.logic.service.sys.SysDomainService;
 public class SysDomainAdminController extends AbstractController {
     @Autowired
     private SysDomainService service;
-    
-    private String[] ignoreProperties = new String[] { "id" };
 
     @RequestMapping("save")
-    public String save(SysDomain entity, HttpServletRequest request, HttpSession session, ModelMap model) {
+    public String save(SysDomain entity, String domainName, HttpServletRequest request, HttpSession session, ModelMap model) {
         SysSite site = getSite(request);
-        if (notEmpty(entity.getId())) {
-            if (!entity.getName().equals(service.getEntity(entity.getId()).getName())
-                    && verifyHasExist("domain", service.getEntity(entity.getName()), model)) {
+        if (verifyCustom("noright", !siteComponent.isMaster(site.getId()), model)) {
+            return TEMPLATE_ERROR;
+        }
+        if (notEmpty(domainName)) {
+            if (!entity.getName().equals(domainName) && verifyHasExist("domain", service.getEntity(entity.getName()), model)) {
                 return TEMPLATE_ERROR;
             }
-            entity = service.update(entity.getId(), entity, ignoreProperties);
+            entity = service.update(entity.getName(), entity);
             if (notEmpty(entity)) {
                 logOperateService.save(
                         new LogOperate(site.getId(), getAdminFromSession(session).getId(), LogLoginService.CHANNEL_WEB_MANAGER,
@@ -50,23 +52,44 @@ public class SysDomainAdminController extends AbstractController {
                 entity.setSiteId(site.getId());
             }
             service.save(entity);
-            logOperateService
-                    .save(new LogOperate(site.getId(), getAdminFromSession(session).getId(), LogLoginService.CHANNEL_WEB_MANAGER,
-                            "save.domain", getIpAddress(request), getDate(), getString(entity)));
+            logOperateService.save(new LogOperate(site.getId(), getAdminFromSession(session).getId(),
+                    LogLoginService.CHANNEL_WEB_MANAGER, "save.domain", getIpAddress(request), getDate(), getString(entity)));
         }
         siteComponent.clear();
         return TEMPLATE_DONE;
     }
 
+    @RequestMapping("virify")
+    @ResponseBody
+    public boolean virify(String name, String domainName, Integer id, ModelMap model) {
+        if (notEmpty(name)) {
+            if (notEmpty(id) && !name.equals(service.getEntity(id).getName())
+                    && verifyHasExist("domain", service.getEntity(name), model)
+                    || empty(id) && verifyHasExist("domain", service.getEntity(name), model)) {
+                return false;
+            }
+        }
+        if (notEmpty(domainName)) {
+            if (notEmpty(id) && !domainName.equals(service.getEntity(id).getName())
+                    && verifyHasExist("domain", service.getEntity(domainName), model)
+                    || empty(id) && verifyHasExist("domain", service.getEntity(domainName), model)) {
+                return false;
+            }
+        }
+        return true;
+    }
+
     @RequestMapping("delete")
-    public String delete(Integer id, HttpServletRequest request, HttpSession session) {
+    public String delete(Integer id, HttpServletRequest request, HttpSession session, ModelMap model) {
+        SysSite site = getSite(request);
+        if (verifyCustom("noright", !siteComponent.isMaster(site.getId()), model)) {
+            return TEMPLATE_ERROR;
+        }
         SysDomain entity = service.getEntity(id);
         if (notEmpty(entity)) {
             service.delete(id);
-            SysSite site = getSite(request);
-            logOperateService
-                    .save(new LogOperate(site.getId(), getAdminFromSession(session).getId(), LogLoginService.CHANNEL_WEB_MANAGER,
-                            "delete.domain", getIpAddress(request), getDate(), getString(entity)));
+            logOperateService.save(new LogOperate(site.getId(), getAdminFromSession(session).getId(),
+                    LogLoginService.CHANNEL_WEB_MANAGER, "delete.domain", getIpAddress(request), getDate(), getString(entity)));
         }
         siteComponent.clear();
         return TEMPLATE_DONE;
